Use [TestCase] and Tidy Tests
Put all github tests into one file
